+++ test/Sdk.Generator.Tests/FunctionExecutor/NotGeneratedTests.cs
@@ -3,6 +3,7 @@
 
 using System.Reflection;
 using System.Threading.Tasks;
+using Microsoft.Azure.Functions.Sdk.Generator.Tests;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Sdk.Generators;
 using Microsoft.CodeAnalysis.CSharp;
@@ -12,7 +13,7 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Xunit;
 
-namespace Microsoft.Azure.Functions.SdkGeneratorTests
+namespace Microsoft.Azure.Functions.Sdk.Generator.FunctionExecutor.Tests
 {
     public partial class FunctionExecutorGeneratorTests
     {