+++ extensions/Worker.Extensions.RabbitMQ/release_notes.md
@@ -4,6 +4,6 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.RabbitMQ 2.0.4
+### Microsoft.Azure.Functions.Worker.Extensions.RabbitMQ <version>
 
-- Update `Microsoft.Azure.WebJobs.Extensions.RabbitMQ` to `2.0.4` (#3019)
+- <event>